feat: various improvements on 'gnodev test'
